Seperated all the fuctions with single task rule
