+++ src/invoiceSections/invoiceHeader.ts
@@ -1,13 +1,12 @@
 import { getStyles } from "./invoiceStyling";
 
 export function getInvoiceHeader(): string {
-    return `<!DOCTYPE html>
+  return `<!DOCTYPE html>
   <html lang="en">
   <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Invoice</title>
   <style>${getStyles()}</style>
-  </head>
-  `;
-  }
+  </head>`;
+}
